[BACKPORT 2.6][PLAT-2613]: Adding a runtime config flag to retain shellprocess cmdOutputLogs.
Summary:
Adding a runtime config flag to allow user to retain shell process commannd ouput logs.
Default value of the flag is true i.e. logs will be deleted until user set the flag to false.

Original commit/Diff:
9c47d7d1bfb24a157672a92879488778eefda09b / D14880

Test Plan:
Verified locally that logs files were retained when flag was set to false.
Jenkins: rebase: 2.6

Reviewers: sb-yb, achauhan, nbhatia

Reviewed By: achauhan, nbhatia

Subscribers: yugaware, jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D14913

+++ managed/src/main/java/com/yugabyte/yw/common/ShellProcessHandler.java
@@ -14,8 +14,8 @@ import com.google.inject.Inject;
 import com.google.common.base.Joiner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import javax.inject.Singleton;
+import com.yugabyte.yw.common.config.RuntimeConfigFactory;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
@@ -34,6 +34,10 @@ public class ShellProcessHandler {
   public static final Logger LOG = LoggerFactory.getLogger(ShellProcessHandler.class);
 
   @Inject play.Configuration appConfig;
+
+  @Inject private RuntimeConfigFactory runtimeConfigFactory;
+
+  static final String COMMAND_OUTPUT_LOGS_DELETE = "yb.logs.cmdOutputDelete";
 
   public ShellResponse run(
       List<String> command, Map<String, String> extraEnvVars, boolean logCmdOutput) {
@@ -121,11 +125,13 @@ public class ShellProcessHandler {
           response.description,
           status,
           response.durationMs);
-      if (tempOutputFile != null && tempOutputFile.exists()) {
-        tempOutputFile.delete();
-      }
-      if (tempErrorFile != null && tempErrorFile.exists()) {
-        tempErrorFile.delete();
+      if (runtimeConfigFactory.globalRuntimeConf().getBoolean(COMMAND_OUTPUT_LOGS_DELETE)) {
+        if (tempOutputFile != null && tempOutputFile.exists()) {
+          tempOutputFile.delete();
+        }
+        if (tempErrorFile != null && tempErrorFile.exists()) {
+          tempErrorFile.delete();
+        }
       }
     }
